Moving cx libraries to zoo directory
